Add files via upload
Two resources are failing to deploy:

 1. jac3sukjdrxzi-apim/zbckw6q67sgh2-func-crewtaskmgrauthsvcs-key This is a named value that tries and fails to reference a property that contains the value (the function key). The error is {"code":"ValidationError","message":"One or more fields contain incorrect values:","details":[{"code":"ValidationError","target":"value","message":"Either Value or Keyvault must be provided."}]}. The property containing the value is on line 1072. The original is commented out on line 989 as a nested child resource and I tried moving it to line 1072 and this did not help.  You can see in the comments where I was originally fetching and storing the function in the variable defaultHostKey and when that was not working tried storing the function key as direct literal and that did not help.

It could also be the case that it is finding the property and the property does not have the correct value for the function key. I noticed that the value of the key generated by exporting the ARM JSON of the APIM seems to have an unusual syntax (it uses forward slashes and I have not seen this in a function key before) and it does not match the value of the default key for the function I see in the portal. When I tried replacing the value of the key from the APIM JSON export with the value I see in the portal for the "Get function URL" for the azure function, this did not help. Same errors.

2. jac3sukjdrxzi-apim/zbckw6q67sgh2-func-crewtaskmgrauthsvcs. This is the back end defined on line 935 that references the azure function and the property (perhaps indirectly via the named-value) that constains the key for the function. I believe this is failing because the name value is failing.

Here is the error: {"code":"ValidationError","message":"One or more fields contain incorrect values:","details":[{"code":"ValidationError","target":"credentials","message":"Property 'zbckw6q67sgh2-func-crewtaskmgrauthsvcs-key' not found."}]}